# artifacts/api-server/src/routes/ratings.ts
import { Router, type IRouter } from "express";
import { db, ratingsTable, ticketsTable } from "@workspace/db";
import { eq } from "drizzle-orm";
import { RateTicketBody } from "@workspace/api-zod";
import { requireAuth, requireActive, requireRoles } from "../middlewares/auth";

const router: IRouter = Router();

router.get("/tickets/:ticketId/rating", requireAuth, requireActive, async (req, res): Promise<void> => {
  const raw = Array.isArray(req.params.ticketId) ? req.params.ticketId[0] : req.params.ticketId;
  const ticketId = parseInt(raw, 10);

  if (isNaN(ticketId)) {
    res.status(400).json({ error: "ID inválido" });
    return;
  }

  const [rating] = await db.select().from(ratingsTable).where(eq(ratingsTable.ticketId, ticketId));
  if (!rating) {
    res.status(404).json({ error: "Avaliação não encontrada" });
    return;
  }

  res.json(rating);
});

router.post("/tickets/:ticketId/rating", requireAuth, requireActive, requireRoles("USER"), async (req, res): Promise<void> => {
  const user = req.user!;
  const raw = Array.isArray(req.params.ticketId) ? req.params.ticketId[0] : req.params.ticketId;
  const ticketId = parseInt(raw, 10);

  if (isNaN(ticketId)) {
    res.status(400).json({ error: "ID inválido" });
    return;
  }

  const [ticket] = await db.select().from(ticketsTable).where(eq(ticketsTable.id, ticketId));
  if (!ticket) {
    res.status(404).json({ error: "Chamado não encontrado" });
    return;
  }
  if (ticket.status !== "RESOLVED" && ticket.status !== "CLOSED") {
    res.status(400).json({ error: "Chamado deve estar resolvido para ser avaliado" });
    return;
  }
  if (ticket.createdById !== user.userId) {
    res.status(403).json({ error: "Apenas o criador do chamado pode avaliá-lo" });
    return;
  }

  const parsed = RateTicketBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const [rating] = await db.insert(ratingsTable).values({
    ticketId,
    userId: user.userId,
    score: parsed.data.score,
    feedback: parsed.data.feedback ?? null,
  }).returning();

  res.status(201).json(rating);
});

export default router;