# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API do Sistema de Suporte TI
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  uf: string;
  municipality: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  USER: "USER",
  COORDINATOR: "COORDINATOR",
  ANALYST: "ANALYST",
  ADMIN: "ADMIN",
} as const;

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const UserStatus = {
  PENDING: "PENDING",
  ACTIVE: "ACTIVE",
  INACTIVE: "INACTIVE",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  status: UserStatus;
  uf: string;
  municipality: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UpdateUserBody {
  name?: string;
  uf?: string;
  municipality?: string;
}

export type ApproveUserBodyRole =
  (typeof ApproveUserBodyRole)[keyof typeof ApproveUserBodyRole];

export const ApproveUserBodyRole = {
  USER: "USER",
  COORDINATOR: "COORDINATOR",
  ANALYST: "ANALYST",
  ADMIN: "ADMIN",
} as const;

export interface ApproveUserBody {
  role: ApproveUserBodyRole;
}

export type TicketType = (typeof TicketType)[keyof typeof TicketType];

export const TicketType = {
  SOFTWARE: "SOFTWARE",
  HARDWARE: "HARDWARE",
} as const;

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus];

export const TicketStatus = {
  OPEN: "OPEN",
  IN_PROGRESS: "IN_PROGRESS",
  RESOLVED: "RESOLVED",
  CLOSED: "CLOSED",
} as const;

export type TicketPriority =
  (typeof TicketPriority)[keyof typeof TicketPriority];

export const TicketPriority = {
  LOW: "LOW",
  MEDIUM: "MEDIUM",
  HIGH: "HIGH",
} as const;

export type UserRefRole = (typeof UserRefRole)[keyof typeof UserRefRole];

export const UserRefRole = {
  USER: "USER",
  COORDINATOR: "COORDINATOR",
  ANALYST: "ANALYST",
  ADMIN: "ADMIN",
} as const;

export interface UserRef {
  id: number;
  name: string;
  email: string;
  role: UserRefRole;
}

export interface Ticket {
  id: number;
  title: string;
  description: string;
  type: TicketType;
  status: TicketStatus;
  priority: TicketPriority;
  uf: string;
  municipality: string;
  createdById: number;
  /** @nullable */
  assignedToId: number | null;
  createdAt: string;
  updatedAt: string;
  createdBy: UserRef;
  assignedTo?: UserRef | null;
}

export interface Message {
  id: number;
  ticketId: number;
  senderId: number;
  message: string;
  createdAt: string;
  sender: UserRef;
}

export interface Rating {
  id: number;
  ticketId: number;
  userId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  score: number;
  /** @nullable */
  feedback?: string | null;
  createdAt: string;
}

export type TicketDetail = Ticket & {
  messages?: Message[];
  rating?: Rating | null;
};

export type CreateTicketBodyType =
  (typeof CreateTicketBodyType)[keyof typeof CreateTicketBodyType];

export const CreateTicketBodyType = {
  SOFTWARE: "SOFTWARE",
  HARDWARE: "HARDWARE",
} as const;

export type CreateTicketBodyPriority =
  (typeof CreateTicketBodyPriority)[keyof typeof CreateTicketBodyPriority];

export const CreateTicketBodyPriority = {
  LOW: "LOW",
  MEDIUM: "MEDIUM",
  HIGH: "HIGH",
} as const;

export interface CreateTicketBody {
  title: string;
  description: string;
  type: CreateTicketBodyType;
  priority: CreateTicketBodyPriority;
}

export type UpdateTicketBodyStatus =
  (typeof UpdateTicketBodyStatus)[keyof typeof UpdateTicketBodyStatus];

export const UpdateTicketBodyStatus = {
  OPEN: "OPEN",
  IN_PROGRESS: "IN_PROGRESS",
  RESOLVED: "RESOLVED",
  CLOSED: "CLOSED",
} as const;

export type UpdateTicketBodyPriority =
  (typeof UpdateTicketBodyPriority)[keyof typeof UpdateTicketBodyPriority];

export const UpdateTicketBodyPriority = {
  LOW: "LOW",
  MEDIUM: "MEDIUM",
  HIGH: "HIGH",
} as const;

export interface UpdateTicketBody {
  title?: string;
  description?: string;
  status?: UpdateTicketBodyStatus;
  priority?: UpdateTicketBodyPriority;
}

export interface AssignTicketBody {
  assignedToId: number;
}

export interface CreateMessageBody {
  message: string;
}

export interface CreateRatingBody {
  /**
   * @minimum 1
   * @maximum 5
   */
  score: number;
  feedback?: string;
}

export interface ReportSummary {
  totalTickets: number;
  openTickets: number;
  inProgressTickets: number;
  resolvedTickets: number;
  closedTickets: number;
  totalUsers: number;
  pendingUsers: number;
  /** @nullable */
  avgResolutionHours: number | null;
  /** @nullable */
  avgRating: number | null;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface RegionCount {
  uf: string;
  count: number;
}

export interface TypeCount {
  type: string;
  count: number;
}

export interface ActivityItem {
  id: number;
  action: string;
  ticketId: number;
  ticketTitle: string;
  userName: string;
  createdAt: string;
}

export type ListUsersParams = {
  status?: ListUsersStatus;
  role?: ListUsersRole;
};

export type ListUsersStatus =
  (typeof ListUsersStatus)[keyof typeof ListUsersStatus];

export const ListUsersStatus = {
  PENDING: "PENDING",
  ACTIVE: "ACTIVE",
  INACTIVE: "INACTIVE",
} as const;

export type ListUsersRole = (typeof ListUsersRole)[keyof typeof ListUsersRole];

export const ListUsersRole = {
  USER: "USER",
  COORDINATOR: "COORDINATOR",
  ANALYST: "ANALYST",
  ADMIN: "ADMIN",
} as const;

export type ListTicketsParams = {
  status?: ListTicketsStatus;
  type?: ListTicketsType;
  priority?: ListTicketsPriority;
  uf?: string;
  municipality?: string;
};

export type ListTicketsStatus =
  (typeof ListTicketsStatus)[keyof typeof ListTicketsStatus];

export const ListTicketsStatus = {
  OPEN: "OPEN",
  IN_PROGRESS: "IN_PROGRESS",
  RESOLVED: "RESOLVED",
  CLOSED: "CLOSED",
} as const;

export type ListTicketsType =
  (typeof ListTicketsType)[keyof typeof ListTicketsType];

export const ListTicketsType = {
  SOFTWARE: "SOFTWARE",
  HARDWARE: "HARDWARE",
} as const;

export type ListTicketsPriority =
  (typeof ListTicketsPriority)[keyof typeof ListTicketsPriority];

export const ListTicketsPriority = {
  LOW: "LOW",
  MEDIUM: "MEDIUM",
  HIGH: "HIGH",
} as const;