# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API do Sistema de Suporte TI
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Registrar novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: E-mail já cadastrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Autenticar usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Autenticado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Perfil do usuário autenticado
      responses:
        "200":
          description: Dados do usuário autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ───────── USERS ─────────
...
  /tickets/{ticketId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: Listar mensagens de um chamado
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lista de mensagens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: createMessage
      tags: [messages]
      summary: Enviar mensagem em um chamado
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageBody"
      responses:
        "201":
          description: Mensagem enviada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ───────── RATINGS ─────────
  /tickets/{ticketId}/rating:
    get:
      operationId: getTicketRating
      tags: [ratings]
      summary: Obter avaliação de um chamado
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Avaliação encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"
        "404":
          description: Avaliação não encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: rateTicket
      tags: [ratings]
      summary: Avaliar chamado resolvido
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRatingBody"
      responses:
        "201":
          description: Avaliação registrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ───────── REPORTS / WOW ─────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        uf:
          type: string
        municipality:
          type: string
      required:
        - name
        - email
        - password
        - uf
        - municipality
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    # ─── User ───
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [USER, COORDINATOR, ANALYST, ADMIN]
        status:
          type: string
          enum: [PENDING, ACTIVE, INACTIVE]
        uf:
          type: string
        municipality:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - status
        - uf
        - municipality
        - createdAt
...
    UserRef:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [USER, COORDINATOR, ANALYST, ADMIN]
      required:
        - id
        - name
        - email
        - role
...
    Message:
      type: object
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        senderId:
          type: integer
        message:
          type: string
        createdAt:
          type: string
          format: date-time
        sender:
          $ref: "#/components/schemas/UserRef"
      required:
        - id
        - ticketId
        - senderId
        - message
        - createdAt
        - sender
    CreateMessageBody:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ─── Rating ───
    Rating:
      type: object
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        userId:
          type: integer
        score:
          type: integer
          minimum: 1
          maximum: 5
        feedback:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - ticketId
        - userId
        - score
        - createdAt
    CreateRatingBody:
      type: object
      properties:
        score:
          type: integer
          minimum: 1
          maximum: 5
        feedback:
          type: string
      required:
        - score

    # ─── Reports ───
...