# lib/api-zod/src/generated/types/createRatingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API do Sistema de Suporte TI
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRatingBody {
  /**
   * @minimum 1
   * @maximum 5
   */
  score: number;
  feedback?: string;
}