# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API do Sistema de Suporte TI
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";
import type { UserStatus } from "./userStatus";

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  status: UserStatus;
  uf: string;
  municipality: string;
  createdAt: Date;
}